requirements.txt
selene==2.0.0rc9
pytest==8.4.1
allure-pytest==2.9.45
allure-python-commons==2.9.45
selenium
python-dotenv